New version
Added a bunch of performance improvements. Still a long way from being optimial however, due to bug in numpypy's rotation code taking a huge amount of time. Presumably numpypy actually rotates the struct, while regular numpy just stores a new view of the struct. Thus, i think i can get numpypy to do the same using a view when writing the data to disk, however, previous tests have shown that plain CFFI C structs are orders of magnitude faster than numpy or numpypy when it comes to simply writing data to a struct (and not performing operations on it), and so a much better path forwards is to remove the numpy dependancy all together, and stick to raw CFFI.